+++ src/Grid/Cell.h
@@ -10,15 +10,15 @@
 
 class Cell {
 public:
-    std::vector<std::vector<float>> S;
-    std::vector<float> A;
-    std::vector<float> Ax;
-    std::vector<float> Ay;
-    std::vector<float> Az;
+    std::vector<std::vector<double>> S;
+    std::vector<double> A;
+    std::vector<double> Ax;
+    std::vector<double> Ay;
+    std::vector<double> Az;
 
 
-    Cell(std::vector<float> Si_, std::vector<float> Sj_, std::vector<float> Sk_);
-    Cell(): A{std::vector<float>(3)}, Ax{std::vector<float>(3)}, Ay{std::vector<float>(3)}, Az{std::vector<float>(3)}{};
+    Cell(std::vector<double> Si_, std::vector<double> Sj_, std::vector<double> Sk_);
+    Cell(): A{std::vector<double>(3)}, Ax{std::vector<double>(3)}, Ay{std::vector<double>(3)}, Az{std::vector<double>(3)}{};
 };
