+++ src/Grid/Grid.h
@@ -21,13 +21,13 @@ private:
 
 
     // Abstraction for grid volume and surface vector calculations.
-    static std::vector<float> calculate_face_vector(std::vector<int>& v1, std::vector<int>& v2,
+    static std::vector<double> calculate_face_vector(std::vector<int>& v1, std::vector<int>& v2,
                                                     std::vector<int>& v3, std::vector<int>& v4,
                                                     Block& b);
     void calculate_block_face_vectors(Block& b);
     void calculate_block_volumes(Block& b);
     static std::vector<std::vector<int>> get_vertex_ijk_vectors(int faceId, int i, int j, int k);
-    std::vector<float> get_face_midpoint_vector(int faceId, int i, int j, int k, Block& b);
+    std::vector<double> get_face_midpoint_vector(int faceId, int i, int j, int k, Block& b);
     void find_grid_min_volume();
 
 public:
